Frontend/requirements.txt
albumentations==1.3.1
builder==0.1.0
cython_bbox==0.1.3
dataloader==2.0
dataset==1.6.2
detectron2==0.6
filterpy==1.4.5
h5py==3.8.0
lap==0.4.0
loguru==0.7.0
matcher==0.2
matplotlib==3.4.2
mmdet==3.1.0
model==0.6.0
motmetrics==1.4.0
numpy==1.23.5
onnx==1.13.1
onnx_graphsurgeon==0.0.1.dev5
onnxruntime==1.15.1
onnxruntime_gpu==1.14.1
onnxsim==0.4.33
opencv_contrib_python==4.8.0.74
opencv_contrib_python_headless==4.8.0.74
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
pafy==0.5.5
pandas==1.5.3
Pillow==9.3.0
Pillow==10.0.0
psutil==5.9.4
pycocotools==2.0.6
PyYAML==6.0
PyYAML==6.0
Requests==2.31.0
scikit_learn==1.1.3
scipy==1.11.1
seaborn==0.12.2
segmentation==0.2.2
setuptools==65.6.3
skimage==0.0
streamlit==1.24.1
tabulate==0.9.0
tensorrt==8.6.1
thop==0.1.1.post2209072238
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.64.1
ultralytics==8.0.125
util==0.0.1
wandb==0.15.5
~pencv_python_headless==4.5.3.56
